requirements.txt
streamlit
scikit-learn
regex
PyPDF2
python-docx
matplotlib
